requirements.txt
pyyaml==5.1.2
libvirt-python==5.7.0
flask==1.1.1
shortuuid==0.5.0
